Dashboard: Add dialog to make the user enter missing contexts
When building context sensitive dashboards, like a dashboard about a
single Checkmk server, the dashboard is built in a way where the
hostname is not hard coded in the dashboard definition. Instead of that
the hostname (context) needs to be added from the outside (via HTTP
parameter).

To make the feature more usable the missing information is now detected
and a popup dialog asking for the missing information (hostname) is
initially shown to the user. After the user completed the information,
the dashboard is being shown as expected.

A menu entry to re-open the popup has been added to the dashboard menu.
This makes it possible to change the context after the initial
rendering.

Change-Id: If27c494188abaa5eb87dfc78f56dfb92e642d475