enhance "mounts" check to have a configurable list of ignored options
The check plugin "mounts" uses a hardcoded list of mount options that
are ignored when comparing the mount options determined at discovery time
(expected options) to the actual mount options during a check. This list
grew over time as new options turned out to be flipping from ON to OFF
for various reasons like energy saving algorithms, for instance.

This commit adds the possibility to configure that "ignore list" by means
of a rule. The old rule allowed to define a list of expected mount options
(overwriting the list that was dynamically determined at discovery time).

The new rule additionaly allows for the definition of an "ignore
list". Both lists are now combined in a dictionary so that the "params"
parameter to the check function has changed from [] to {} and was renamed
from "targetopts" to "params".

The check and the WATO plugin are both compatible with existing rules.