Add query for vmware fault tolerance mode
If VMWare fault tolerance mode is enabled vmware creates and maintains
another identical machine which is continuously avaible and replace it
with the primary virtualmachine in case of a failover situation.
The secondary VM still shows up in the agent output with the name filled
with zeros on cpu / memory. Depends what was queried first the service
output was filled with the correct data or zeros.
This is fixed by getting the ftrole and dont write the information of
the secondary vm into the agent output.

Close #208

Change-Id: I5989b7010706f88ed5f2b5cdcf314196ba9b64a5