The inline method for popups now only makes popups visible
The old inline method used a string representation of
HTML stored in the onclick handler of the link that
opens the popup.

Now the popups are expected to be part of the DOM and
can be made visible solely with CSS.

Although the method is not used yet, it will be used
later on in the UX project.

Change-Id: Icd2a16e7511f414ca1e45a6615515174952c4622