snmp_scan: Refactoring #2 - pass sections later
The list of sections is computed as late as possible, that is, in the
CachedSNMPDetector.  The scan function does not need to know whether the
caller is discovery or inventory.  It only needs the corresponding
SNMPSectionPlugin iterable.

This removes more deps of snmp_scan on base.

We can now unit test the module.

CMK-4766

Change-Id: I0df39233aae36a1944b831abddd421e358929832