CheckAPI: make SectionName a PluginName
SectionName is no longer a alias for str, but a proper class instead.
This allows to enforce a limited character set, and a better (enforced)
distiction between other classes of ABCPluginNames.

Dumbing down raw_data in data_sources/snmp.py for writing to
cache file is required.

This also resolves CMK-4461

Change-Id: If880a3ac9581a58fb0100f4103bf2a2418113da3