Display pylint version in the pylint test
It is not uncommon that new versions of pylint discover new linting
errors.  However, this may come a little unpleasant when the new errors
cannot be related to their cause: the new pylint version.

This patch makes sure that the version of pylint used in the tests is
shown in the tests.

Change-Id: I75425b507619623e9ac4ade42acd5e6c1d42d8f6