CheckAPI: make autochecks accept Service instances
The DiscoverdService object mainly deals with the problem of the unresolved parameters potentially read from the autochecks file and returned by the discovery function. Since we want to get rid of these unresolved parameters (the former case in the long run, the latter case is already history), we allow to write "fully qualified" Service-instances to the autochecks file.
When reading the file  we must assume they are only DiscoveredService instances, unforunately.

Change-Id: I53c01ba816ef54b3eff2a8de394180fbe3064658