config: add option to migrate check plugin variables
When resolving the parameters saved in the autochecks file, we may need
certain variables to be defined in the check context and in config.get_check_varisbles().

When migrating a check plugin to the new check API the corresponding check context
will not be present any more. You can add those variables, in order to be able
to resolve the variables in saved autochecks.

This will also have the effect, that variables defined in a *.mk file in etc/check_mk/conf.d
will be considered.

For example: When migrating the check plugin `icom_repeater` we must add the following line
to cmk.utils.migrated_check_variables, in order to be able to read and parse the autochecks
that make use of such a variable (i.e. discovered pre checkmk 1.7):

icom_ps_volt_default_levels = (13.5, 13.2, 14.1, 14.4)

Change-Id: I3cd2a05a96d70374c7fd8ec024fed1a40df8701c