11045 Optional package: Replace Kubernetes extension package
This change replaces the optional extension package
k8s_ingress_and_job_info for Kubernetes by the extension
package k8s_extensions.

It fixes the bug that the Kubernetes Ingresses and Jobs
could not be selected in the datasource ruleset of the
former extenion package.

Furthermore the ruleset now contains an option to use a
namespace prefix for the hosts that are created in
Checkmk and Endpoints can now be monitored.

Change-Id: Ib799f2bf6ba13879bba1ace4317979a6eff76b4c