Reorganize data source modules
`DataSource` and `HostSections` classes work in pair.  We make that
clear by grouping:

 - AgentHostSections and AgentDataSource (was CheckMKAgentDataSource)
 - SNMPHostSections and ABCSNMPDataSource (were already together)
 - AbstractHostSections and DataSource (moved together to a new abstract
 module)

We now obtain better symmetry and split modules by abstraction level.
This better underlines the structure of the code.

We move concrete classes and free functions out of `abstract`.  The name
of the module really *implies* that it contains abstract classes and
interfaces.

 - Utility classes needed for caching move to `_cache`
 - Utility free functions move to `_utils`

We empty `__init__.py`.  It is difficult to avoid circular dependencies
with code in `__init__`.  Now that DataSources is in its own module, we
have full control on the names at package scope.

The callers can now use module names as namesapce, which is rather
common and expected in Python.

CMK-4726

Change-Id: Ic993887df194552f6c5fdfdc4c2bb56d76b72ae0