SRP: Extract SourceBuilder class
A large number of DataSources methods serve to make its `sources`
attribute in `DataSources.__init__()`.  They are therefore not part of
the DataSources object.  We move these methods in a SourceBuilder class
and set `DataSources.sources` with a free function.

We further add a trivial `make_description` that removes one more method
from the class.

CMK-4726

Change-Id: Ifde41c9a4797fa7db8310aa9160f546807685c5b