CheckAPI: clean up parameter handling
Legacy check plugins could have all kinds of parameters as default paramters or discovery
results.
In the new API we validate that only a dictionary must be used, and we pass a read-only variant
of it to the check function (TODO: and the discovery function). This requires a hack to wrap/unwrap
the legacy parameters, to pass the validation.

Change-Id: Ia3cd96c3def9b3cf01324c1abc75f39af23a6fcc