11221 FIX: systemtime: Offset for vSphere special agent
The check <tt>systemtime<\tt> reported an incorrect offset when checking
the output of the vSphere special agent. This was caused by the runtime of
the special agent itself and has now been fixed. The check <tt>systemtime<\tt>
can now process a second, optional input which gives the timestamp against
which the actual offset is computed.

Change-Id: I23d6650e5bfb56fd109d8f3e6997aefab8c39bba