Make MHS API more consistent
Now that we have established that MHS is a mapping, we consistently pass
the key as the first argument to its methods.  The added consistency
simplifies the class and shows its purpose better.

The remaining methods that do not take HostKey are most likely not where
they should be.  We make this clear by making them static and by
grouping them at the end of the class definition

CMK-4838

Change-Id: I0602b7611394b29601970fa6d22a610d620ed45c