Do not pass host config to datasources
The DataSources class does not really need it.  As a result we must pass
the list of sources to the DataSources container.  This is a rather
common pattern, indeed, the responsibility of containers is to contain,
not to fill itself magically.  Callers then pass the sources.

Passing the list of sources further increases the overall testability of
the class.  It  further makes it clear that
`DataSources.get_host_sections()` barely touches DataSources.

We also move the agent description to HostConfig.  It had nothing to do
with DataSources and everything with HostConfig.

CMK-4838

Change-Id: I1cea1810f360cf24293b7875a69e9e3f195eaffc