Make get_host_sections static
Instead of playing guess-games with default arguments in the function,
we let the caller pass the good values.

This results in code that is much simpler and further shows that
`get_host_sections` needs not be in DataSources.  Indeed,
`DataSources.get_host_sections()` returns a MultiHostSections.  But
MultiHostSections is but a glorified `Mapping[HostKey,
AbstractHostSections]` and the arguments to `get_host_sections()` could
be easily rewritten as `[HostKey, ...]`.

This is the way to close CMK-4838.

CMK-4838

Change-Id: Ie3eae73ad4dedce4e00bd7474ca71abb1891c200