Python 3: Remove compatibility automations
Re-enable in GUI process rule match testing and description computation
for performance reasons. The automation approach is a better separation,
but very bad for the performance of those pages.

We'll have to find a better solution in the future. This commit reverts
the py2/3 compatibility workaround to the state we had with 1.6.

Change-Id: I26c3c3dd8a540efca185fbda375fdeeb724d236a