11139 FIX ESX inventory plugin
The inventory plugin "esx_systeminfo" was fixed and adapted to the parsed checkoutput.
If you do not see your ESX host in the inventory, perform a new Check_MK HW/SW Inventory on the relevant host after applying this fix and the ESX inventory should appear.
Thanks to Christian Dewald for the bug fix!

Change-Id: If2f028166939457cf3c9475f9b193eb3cf5026b7