Change host section initialization back to old logic
We are not completely sure whether or not we already have sections in the
MultiHostSections object. We must either get the existing one or create a
new one and add it to MultiHostSections.

To archieve this, we use setdefault, which is similar to the approach we
used in 1.6, but with a bit cleaner MultiHostSections API.

The real time checks can be sure that they don't need to respect existing
sections and directly set the new section.

Change-Id: I73f4a1971b2e8fea01db09b13dfa58733825e918