11248 Agent bakery: Automatically build agents for created hosts
After creating a new host in the configuration, the agent baking process for
the new host is triggered now. When multiple hosts are created at once, for
example using the API or an import mechanism, one baking process is started for
all of these hosts.

The process is started as a background job, just like when a bake process is
started manually. With the single difference that the process only cares about
the newly added hosts. You can see the process or result message when accessing
the Agent Bakery page in your Configuration.

Change-Id: I342d56a4e4598816247f6f0088def7b323784e01