11275 FIX ps - Size of processes: Change graph visual display order
For graphs of "ps" containing information about "Size of processes" the visual orderning of the graphs is fixed.
Before "Resident size" and "Virtual Size" were stacked on top of each other, which was wrong.
Resident size is a subset of virtual size. The data of the graph "Resident size" is now visually in front of the other graphs.

Change-Id: I4ed49dd8ddd73ba7e35a01b846d973ec7a7a4fc2