core config: ignore snmp_'check'_interval
A check plugin is no longer in 1:1 correspondence with an SNMP section.
The variable referred to as snmp_check_interval is not a check interval: It is the timeperiod for which a specified SNMP sections data is cached. It has no influence on the interval with which a check plugin is executed.

Change-Id: I2f97d101904cca9eceba1d3ed7e4b0a9fb0177f0