data_sources/snmp: nuke _for_inventory
The data source does not need to know, wether we are intending to do an inventory or not.
Either we pass a list of 'interesting' raw sections upon instantiation, or we perform an
SNMP scan (and get all raw sections there are).

Wether or not we want to execute inventory plugins should be considered
during the determination of the relevant raw sections names.

CMK-4779

Change-Id: I81600661b544ec1f6e5d7b5c535323e6b9a08db5