10783 Resolve composing metrics in Custom Graphs
When adding all the metrics from a graph into a Custom Graph, the metrics
that where combine through a certain operation would remain as such in the
Custom Graph.

With this werk the composing metrics are resolved, so that the Custom Graph
can dissolve the operation. This is particularly useful when integrating a
lot of metrics from a Combined Graph into the Custom Graph, and posterior
editing is desired.

CMK-4746

Change-Id: I045bdbcefa500971fb69a792f2a180b3aadbf924