Make DataSources private
The instantiation of DataSources receives a list of sources as argument
and iteration then iterates over this list.  We can skip the
instantiation altogether and use the list as is.

Now, we can also avoid passing the list as an argument to functions and
make the list ad-hoc where it is needed.

This patch also orders arguments passed to functions consistently.

CMK-4838

Change-Id: If6ebe4062ad5d420b8fdb0c260cea2425eb546ca