11086 FIX Docker container: Fix usage behind reverse proxies
The Checkmk container was not operating as expected in different scenarios
where it's user interface was accessed through a reverse proxy, for example
when it was working behind a K8s load balancer.

This behavior now has been improved by removing the port from the "ServerName"
directive in the docker container. This makes the web server of the container
detect and use the port from client requests and reuses it to build URLs for
redirects and so on.

Change-Id: I4a61fd8d494d6e818d6a76e71947465b73e036bc