11047 FIX mknotifyd, mknotifyd.connection: resolve wrong warnings
If another process tried to connect to the mknotifyd via TCP the
notification spooler service returned CRIT together with the
message "The state file seems to be empty or corrupted. It is very
likely that the spooler is not working properly". This wrong warning
is now fixed. Furthermore, any non-readable characters in the
message header of log entries starting with "Bogus message
header 'message header'" are now represented as '\xhh' where h is
a hexadecimal number.

Change-Id: I80df4e98f5be549582eb39cf4cdae6352df7249e