Fix core config cache update
When the core config loading was not triggered by get_canonical_name and
get_config_for_host was called first, the cache was not populated and the
EC was working with the default config instead of the config that is
currently active in the core.

This could make the EC miss e.g. host individual limit rules.

Change-Id: I00a7b47947cdc39182990b48b699b06c02266292