Add topic migration for custom visuals
This affects views, reports, dashboards, custom graphs, graph
collections and forecast graphs.

In previous versions the topic of these definitions was just a plain text
field which allowed the user to enter a topic which was then used to
structure the definitions in the navigation.

The new navigation requires an unique ID, a title, icon and sort index.
To be able to deal with custom definitions.

With this migration logic we load all custom definitions from the site
and migrate them to the new mechanic. This is done by first searching
for existing builtin topics that match the topic, then referencing it or
creating a new topic with the same owner as the custom definition. The
custom definitions will then reference to the topics.

This migration is done during site update. All updates will be made
before the GUI is loaded.

Change-Id: I8cb61c90e5e04e1e23778165ccfbef3e0a5c377a