Extract Summarizer class from AgentDS
The AgentDataSource has several responsibilities so that the code is not
clear.  Here, we extract the summary-related methods from the main
class.  It is clear that the methods do not have much to do with the DS.

Now, we can also see that what gets summarized is not the DS but the
AgentSectionContent contained in some attribute of the DS.  We shall
make this clear in a next patch.

CMK-4838

Change-Id: Ib2de63202dd6c7f1e08336e0dbe063363df00db5