Clarify summary object in Agent DS
Pass the Optional[AgentSectionContent] to the summarizer as it
seems like the object being summarized.

We make some trivial methods static.  Actually, the remaining
non-static methods only use an HostConfig that is not really
owned by the Summarizer...  All of this may be advantageously
replaced with free functions or moved some place it belongs in.

CMK-4838

Change-Id: I223f2160a8cd7de4d776e6c129da778e2bf89213