Migration of section and check plugin aix_diskiod, part II & 11223 aix_diskiod: Switch to new diskstat-ruleset
The check <tt>aix_diskiod<\tt> now uses the rules defined under the
ruleset "Levels for disk IO" instead of "Levels on disk IO (old style
checks)". User-defined rules have to be adjusted to account for this.

This commit also migrates the section and the check plugin aix_diskiod
to the new check API.

Change-Id: Ib81a4587bba1a029c0d95da2f612d34c8052fec7