11305 config: drop support of 'precompile_params'
This werk only affects users who have developed their own plugins using the 'precompile function' feature.
If you are not aware of this feature, you are most likely not affected.

The support of this feature is now dropped, so if you have used it, please adapt your code accordingly.
Otherwise your check function will not be passed the expected parameters.
Have a look at the commit history of the sap or fileinfo plugins for reference.

Change-Id: Ib9dbe4c24856c2baeae1ed11878287c609135777