Do not pass SelectedRawSection to DS.__init__
SelectedRawSection are used for filtering either before the fetcher
(SNMP) or after (Agent).  Passing the SelectedRawSection to the DS is,
therefore, not required.  Instead, it make the API more complex than
necessary (see the SourceBuilder for example) and does not underline the
purpose of the parameter.

Moreover, with the new API, the core only passes the hostname to the
Checker, anything else (including the SelectedRawSection) must come
somewhere else.

CMK-4976

Change-Id: Icd0f88677f298538e5abb1bebb7e4716d2b26e55