11224 FIX winperf_phydisk: Interpret averaging parameter correctly
The check <tt>winperf_phydisk<\tt> now interprets the timespan parameter for
computing averaged values correctly (ruleset "Levels on disk IO (old style
checks)"). Previously, the user was supposed to specify a value in minutes,
however, the check plugin actually expected a value in seconds. This has been
corrected.

This change is incompatible, since the check plugin now behaves differently
for a given averaging timespan. User-defined rules might have to be adjusted
to account for this.

This change also migrates the section and the check plugin winperf_phydisk
to the new API.

Change-Id: I6d74b71352fd516a62eefcb748f9b291d445708b