Change algorithm to compute WATO breadcrumb
We now use a recursive approach with the option for each mode
to a) change the way the breadcrumb of itself is built or
b) override the computation of the URL to be linked to.

Especially b) will be useful in case a parent mode also needs some
context for it's link. Like "BI packs > rules > ..." needs it.

Change-Id: I569ae5b2f808edc8af4bad707a992a164ef17227