10127 FIX Do not remove acknowledgements for retries.
When a non-sticky acknowledgement has been set for a host/service in a soft
non-OK state, it has been removed when it stayed in a soft non-OK state in
the next check cycle (i.e. after a retry).  This behavior was unexpected
for the user and inconsistent with the way Nagios behaves, so this has been
fixed.

Change-Id: I343290c422a1525337616cce1bb72a82188062c0