Lift program fetcher config off Program DS
In a first step, we make the configuration steps static methods.  This
way, they can easily be moved around.

We also had to rework the cmdline/stdin test as the existing one could
not handle the call to static methods from __init__.

Regarding the test, we also make a 3 tuple of the input and give names
to both of the "expected" fields.

CMK-5044

Change-Id: I62742a8d25736edfdeacae625643bbe9d2efc7da