11276 FIX jira notification: fix issue creation
Creation of Jira issues sometimes failed because the field 'labels' is not existing.
Jira Service Desk does not have the 'labels' field by default, but Jira Software does.
With this change, the labels field will only be set if the WATO rule for 'labels' is activated.
If 'labels' is activated but not value is defined it will use the default labels 'monitoring'.

Change-Id: I745653448dd193eb231e5a77aae8ab05d421f6a4