add detection for check-state
atm checks turn md-service into warning state, bc. they come with "finish" and "speed" values. Checks are "By default, run at 00:57 on every Sunday, but do nothing unless the day of the month is less than or equal to 7 [...]". 
A check is no rebuild or resync, it's just a periodic check and should degrade the state of the raid.
Discussion about this topic: https://forum.checkmk.com/t/mdadm-check-causing-warning/17813/8