11149 mkp_prometheus: addition of mkp version 1.2
This change updates the optional extension package for the Prometheus suite. The version 1.2
further improves handling of the Kube-state and cAdvisor information. It also allows Checkmk
to correctly handle cases when information is incomplete.
If the Prometheus package v1 was already in use, you should perform a Tabula Rasa after
upgrading to the package v1.2.
For a first installation, you can find this package in the Extension packages page. The
Prometheus package is listed as an optional extension package and becomes available to your
site after performing the "Install this package" action.

Change-Id: I6b75dec922db4f5e09903100b70240aaf823d718