Icons can now be displayed for each item in the mega menu
Icons can now be displayed per item (instead of per
topic) in the mega menu. To activate this display
style a user has to enable the option "Icons per item"
in his personal settings.

Duplicate icons have to be replaced and missing icons
have to be added in upcoming commits.

Change-Id: Ia9f0126dbb8ae097526d41f27b706b701c12db9c