Deobfuscate persisted_sections
`persisted_sections` is used to configure the SNMP fetcher.  It should
therefore move to the SNMPConfigurator.

Analysis of the code shows that:

 1. `persisted_sections` is reset upon entering `ABCDataSource._run()`.
 2. `_run()` calls `_get_raw_data()` and `_get_raw_data()`
 calls `_execute()`.
 3. `persisted_sections` is only really used in
 `SNMPDataSource._make_oid_infos()`
 4. `SNMPDataSource._make_oid_infos()` is only called by
 `SNMPDataSource._execute()`.

Therefore, the instance attribute was never actually used.  Instead, we
pass the `persisted_sections` from `ABCDataSource._run()` to
`SNMPDataSource._make_oid_infos()`.  Having it as an argument instead
of an instance variable should make further modifications easier.

CMK-5044

Change-Id: I45abdb9e29839bf0ed9a87d2939d0761fe8c005a