prepare determination of cache info for 'supersedes' feature
Once the 'supersedes' feature is implemented, any given parsed
section is created by one of various raw sections.
Since we need to parse them in order to see which of them finally
creates the parsed section in question, we must ensure the parsing
has taken place before we can determine the cache info associated
with the parsed section.

In practice we always parse the sections anyway (as we only care for
cache info of check results).

Change-Id: I4e6f5e4840252f5df277708ae2c30e4e61dc6370