Check API: rank parsed section creators by supersedes
Raw sections may get renamed once they are parsed, if they declare it.
The get_parsed_section method of the host sections class now deals with
the task of determining which sections we need to parse, in order to end
up with the desired parsed section.

They are ranked according to their supersedings.

CMK-4232

Change-Id: I097e1ba7d17c5457657bf24389951a354505d9eb