Replace ClassRegistry and InstanceRegistry by Registry
The classes ClassRegistry and InstanceRegistry are replaced
by the class Registry. Registry can handle the registration
of both classes and instances. The docstring of Registry
describes how both types of registries can be defined. All
existing registries are migrated to the new Registry class.

Since mypy can now perform type checks for the objects
stored inside the registries over 70 new typing errors
were revealed. They are fixed or suppressed in this commit
as well to make the CI happy. Errors that are not fixed in
this commit are suppressed with a FIXME comment.

Change-Id: I8e9ebfeaf9da1997f551741a53accd19ade48295