Page menu: Add visual context specific dropdowns
Use the visuals gathered by the previously existing context link
detection for the new context sensitive dropdowns.

To make this possible the visuals just need to be grouped by
the infos they work with to get the visuals for each dropdown
and then by their topic to create the topics for the dropdowns.

Change-Id: Ia8b5736ac09cfa4ef6da906fb3653c007152cd7e