Add availability page menu entry
The entry works just like the availability button before. It opens
the currently visible view in availability mode. This means it shows
the availability for all the rows that are shown to the user on the
resulting page.

The tricky part here is to find out in which of the page menu dropdowns
the link needs to be placed.

Change-Id: Ie4434df41d123c50baece69f6c2ffb4dcddb0266