Fix: Graph metric color matches the Service Metrics painter
The translated metrics deal with generating a metric color, defined by the
metric info or a unique color per set of unregistered metrics.

Graphs use that metric defined color instead of jumping to a new default
for metrics not integrated in the metric system.

The single new default is for scalars(Warn, crit) not defined(from user).

Change-Id: Ie10a227ff31f4f445e5aee38895a1d2dd52f759d