REST-API: activate changes test
add LiveStatus mocking to activate-changes test, because I realized that
requests were being made by the code, the error to connect would not
surface in the code but rather, after the connection timed out, the test
continued to run.

This also explains why the test took so long to run. Now it is actually
covering more code, is more correct and runs faster.

Change-Id: I62a3043acfe2711c143ed411c7619630f5e9bd7b