Move view specific request variable hacks to start of request
This logic only needs to be executed when initially accessing a view,
not e.g. during subsequent AJAX calls. To have a cleaner separation and
have all variables set before the context is computed from the
variables, we move it to the central entry point of the views:
page_view().

Change-Id: Ie3b996380448e02e195706943cce8913a01d980d