11340 md-raid: add detection for check-state
At the moment (md-)checks turn md-service into warning state, because they come with "finish" and "speed" values.
(md-)checks are "By default, run at 00:57 on every Sunday, but do nothing unless the day of the month is less than or equal to 7 [...]".
A (md-)check is no rebuild or resync, it's just a periodic check and should degrade the state of the raid.
Discussion about this topic: https://forum.checkmk.com/t/mdadm-check-causing-warning/17813/8

Thanks to Felix for the Pull Request!
Closes #222

CMK-5058

Change-Id: I254047a0f03445e6774792d7cd9f27008addbcab