Unbreak single group views for host / services
The verification of the single contexts has been introduced some weeks
ago and now prevents rendering of "host of host group X" views. To make
the views usable again the recently introduced verification is being
disabled for this special case.

This is another workaround on top of a collection of workarounds to make
the linking from host group lists to members of single host group page
possible in the current structure.

Change-Id: I52b7e7c311720f66d97da2eb1388cf452ef20382