Generalize switching between popups
When one page menu dropdown was opened it was possible to hover the next
menu title to switch to the other dropdown.  This specific feature has
now been added to the generic popup menu code.

To use this feature the popups need to have a common CSS class which
needs to be added as popup_group="class" argument to the popup_trigger
call in python.

Change-Id: I75ad4a689f13a136972c048542120613d99ee92d