Services: Move toggles to page menu
Because of the dynamic nature of the page content, the page menu currently
also needs to be re-rendered by the AJAX call that updates the page content.
This is somehow possible, but we should consider to refactor this to initially
render all menu entries and show/hide and enable/disable them with every AJAX
update.

The breadcrumb rendering needs to be fixed in the next steps.

Change-Id: Ie1ba6d25600e77ac2d290435f3a61cb77955814b