Configure SNMP scan cache with configurator
In some cases, the `SNMPConfigurator` calls
`CachedSNMPDetector.__call__()` where an SNMP scan is run an memoized.
That means that the configurator may perform unnecessary, slow, and
blocking SNMP I/O.

Instead, we would like to perform I/O only when it is required and also
to localize all I/O in the fetchers.

In a first step, we move the configuration of the detector to the SNMP
configurator so that it may be serialized to the fetcher and that the
SNMP scan may be performed / cached later in the fetcher.

CMK-5100

Change-Id: Iff5ddf6d5f7a391db05fea37bcce1504d36ee8b4